# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Epic Poetry Cafe Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /uom:
    get:
      operationId: listUom
      tags: [uom]
      summary: List units of measure
      responses:
        "200":
          description: UOMs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Uom"
    post:
      operationId: createUom
      tags: [uom]
      summary: Create UOM
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUomRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Uom"
  /uom/{id}:
    patch:
      operationId: updateUom
      tags: [uom]
      summary: Update UOM
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUomRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Uom"
...
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List vendors
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [vendors]
      summary: Create vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVendorRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      summary: Get vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    patch:
      operationId: updateVendor
      tags: [vendors]
      summary: Update vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVendorRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}/spend-summary:
    get:
      operationId: getVendorSpendSummary
      tags: [vendors]
      summary: Vendor spend summary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Spend summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorSpendSummary"
...
  /waste:
    get:
      operationId: listWasteEntries
      tags: [waste]
      summary: List waste entries
      parameters:
        - name: fromDate
          in: query
          schema:
            type: string
        - name: toDate
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Waste entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WasteEntry"
    post:
      operationId: createWasteEntry
      tags: [waste]
      summary: Create waste entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWasteEntryRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WasteEntry"
  /waste/{id}:
    patch:
      operationId: updateWasteEntry
      tags: [waste]
      summary: Update waste entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWasteEntryRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WasteEntry"
  /waste/summary:
    get:
      operationId: getWasteSummary
      tags: [waste]
      summary: Waste analytics summary
      parameters:
        - name: fromDate
          in: query
          schema:
            type: string
        - name: toDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Waste summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WasteSummary"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - fullName
        - role
        - active
    CreateUserRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
        active:
          type: boolean
      required:
        - username
        - password
        - fullName
        - role
    UpdateUserRequest:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
        active:
          type: boolean
        password:
          type: string
...
    Uom:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        abbreviation:
          type: string
        baseUomId:
          type: ["integer", "null"]
        conversionFactor:
          type: number
        active:
          type: boolean
      required:
        - id
        - name
        - abbreviation
        - active
    CreateUomRequest:
      type: object
      properties:
        name:
          type: string
        abbreviation:
          type: string
        baseUomId:
          type: integer
        conversionFactor:
          type: number
        active:
          type: boolean
      required:
        - name
        - abbreviation
...
    Vendor:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        contactPerson:
          type: ["string", "null"]
        mobile:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        gstNumber:
          type: ["string", "null"]
        paymentTerms:
          type: ["string", "null"]
        creditDays:
          type: ["integer", "null"]
        preferred:
          type: boolean
        active:
          type: boolean
        remarks:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - code
        - name
        - preferred
        - active
    CreateVendorRequest:
      type: object
      properties:
        name:
          type: string
        categoryId:
          type: integer
        contactPerson:
          type: string
        mobile:
          type: string
        email:
          type: string
        address:
          type: string
        gstNumber:
          type: string
        paymentTerms:
          type: string
        creditDays:
          type: integer
        preferred:
          type: boolean
        active:
          type: boolean
        remarks:
          type: string
      required:
        - name
    VendorSpendSummary:
      type: object
      properties:
        vendorId:
          type: integer
        totalSpend:
          type: number
        purchaseCount:
          type: integer
        ingredientCount:
          type: integer
        lastPurchaseDate:
          type: ["string", "null"]
      required:
        - vendorId
        - totalSpend
        - purchaseCount
        - ingredientCount
...
    WasteEntry:
      type: object
      properties:
        id:
          type: integer
        wasteNumber:
          type: string
        wasteDate:
          type: string
        wasteType:
          type: string
        ingredientId:
          type: ["integer", "null"]
        ingredientName:
          type: ["string", "null"]
        menuItemId:
          type: ["integer", "null"]
        menuItemName:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        reason:
          type: ["string", "null"]
        quantity:
          type: number
        uom:
          type: string
        costValue:
          type: number
        department:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        approvalStatus:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - wasteNumber
        - wasteDate
        - wasteType
        - quantity
        - uom
        - costValue
        - approvalStatus
    CreateWasteEntryRequest:
      type: object
      properties:
        wasteDate:
          type: string
        wasteType:
          type: string
        ingredientId:
          type: integer
        menuItemId:
          type: integer
        categoryId:
          type: integer
        reason:
          type: string
        quantity:
          type: number
        uom:
          type: string
        department:
          type: string
        notes:
          type: string
      required:
        - wasteDate
        - wasteType
        - quantity
        - uom
    WasteSummary:
      type: object
      properties:
        totalWasteValue:
          type: number
        totalWasteEntries:
          type: integer
        wasteByCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryBreakdown"
        wasteByReason:
          type: array
          items:
            $ref: "#/components/schemas/CategoryBreakdown"
        wasteTrend:
          type: array
          items:
            $ref: "#/components/schemas/TrendDataPoint"
      required:
        - totalWasteValue
        - totalWasteEntries
        - wasteByCategory
        - wasteByReason
        - wasteTrend
...
    TrendDataPoint:
      type: object
      properties:
        date:
          type: string
        value:
          type: number
        label:
          type: ["string", "null"]
      required:
        - date
        - value
    CategoryBreakdown:
      type: object
      properties:
        name:
          type: string
        value:
          type: number
        count:
          type: integer
      required:
        - name
        - value
...